+++ Source/Dafny/DafnyOptions.cs
@@ -59,7 +59,13 @@ namespace Microsoft.Dafny {
       clo = options;
     }
 
+    public enum DiagnosticsFormats {
+      PlainText,
+      JSON,
+    }
+
     public bool UnicodeOutput = false;
+    public DiagnosticsFormats DiagnosticsFormat = DiagnosticsFormats.PlainText;
     public bool DisallowSoundnessCheating = false;
     public int Induction = 4;
     public int InductionHeuristic = 6;
@@ -315,6 +321,26 @@ namespace Microsoft.Dafny {
             int verify = 0;
             if (ps.GetIntArgument(ref verify, 2)) {
               DafnyVerify = verify != 0; // convert to boolean
+            }
+
+            return true;
+          }
+
+        case "diagnosticsFormat": {
+            if (ps.ConfirmArgumentCount(1)) {
+              switch (args[ps.i]) {
+                case "json":
+                  Printer = new DafnyJsonConsolePrinter { Options = this };
+                  DiagnosticsFormat = DiagnosticsFormats.JSON;
+                  break;
+                case "text":
+                  Printer = new DafnyConsolePrinter { Options = this };
+                  DiagnosticsFormat = DiagnosticsFormats.PlainText;
+                  break;
+                case var df:
+                  ps.Error($"Unsupported diagnostic format: '{df}'; expecting one of 'json', 'text'.");
+                  break;
+              }
             }
 
             return true;
@@ -990,6 +1016,10 @@ Exit code: 0 -- success; 1 -- invalid command-line; 2 -- parse or type errors;
 /printTooltips
     Dump additional positional information (displayed as mouse-over tooltips by
     the VS plugin) to stdout as 'Info' messages.
+/diagnosticsFormat:<text|json>
+    Chose how to report errors, warnings, and info messages.
+    text (default): Use human readable output
+    json: Print each message as a JSON object, one per line.
 
 ---- Language feature selection --------------------------------------------